A simple Stream-K kernel with wait-free atomic synchronization and deterministic results (#1)
The synchronization between workgroups cooperating on a tile is
wait-free and produces deterministic results. This is achieved using a
per-tile atomic counter to determine which workgroup is the last to
complete its share of the tile's computation. That workgroup will then
go on to perform the final reduction of the multiple workgroups' local
accumulators into the final accumulator. The results are still
deterministic thanks to performing that final reduction in order of
increasing K-dimension indices, regardless of which workgroup happens to
be doing it.

This PR also includes a necessary refactoring to allow for Stream-K
kernels and to allow for kernels to request an auxiliary device buffer.

This PR also fixes an existing race condition in an unrelated kernel
(`MmtKernel_256t_MSxNS_amdgcn_mfma_f32_16x16x4f32_shared_Kx4_doublebuffer_naive`).
It wasn't a high-performing kernel anyway.

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>