StreamK variant with wait-free binary tree reduction with deterministic results. (#12)
This adds a StreamK kernel where the final reduction is distributed over
as many workgroups as possible by a binary tree reduction.

This is still wait-free and still produces deterministic results,
although the results differ from those produced by a linear reduction.

Performance is anywhere between 0.5x and 1.1x of the linear-reduction
StreamK kernel depending on the shape. That is hard to interpret though
as that is only a matmul microbenchmark, the inner MAC loop is not
optimized, etc. The value of this kernel as it stands is more as a proof
of concept.

---------

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>